Refactoring: move all exceptions in a single module
