+++ resources/lib/playlists/kodi_pl.py
@@ -7,11 +7,13 @@ from __future__ import absolute_import, division, unicode_literals
 from logging import getLogger
 import re
 
-from .common import Playlist, PlaylistError, kodi_playlist_hash
+from .common import Playlist, kodi_playlist_hash
 from . import db, pms
 
 from ..plex_api import API
 from .. import utils, path_ops, variables as v
+from ..exceptions import PlaylistError
+
 ###############################################################################
 LOG = getLogger('PLEX.playlists.kodi_pl')
 REGEX_FILE_NUMBERING = re.compile(r'''_(\d\d)\.\w+$''')
