+++ resources/lib/playlists/plex_pl.py
@@ -6,8 +6,9 @@ Create and delete playlists on the Plex side of things
 from __future__ import absolute_import, division, unicode_literals
 from logging import getLogger
 
-from .common import PlaylistError
 from . import pms, db
+from ..exceptions import PlaylistError
+
 ###############################################################################
 LOG = getLogger('PLEX.playlists.plex_pl')
 # Used for updating Plex playlists due to Kodi changes - Plex playlist
