+++ resources/lib/playlists/pms.py
@@ -7,11 +7,11 @@ manipulate playlists
 from __future__ import absolute_import, division, unicode_literals
 from logging import getLogger
 
-from .common import PlaylistError
-
 from ..plex_api import API
 from ..downloadutils import DownloadUtils as DU
 from .. import utils, app, variables as v
+from ..exceptions import PlaylistError
+
 ###############################################################################
 LOG = getLogger('PLEX.playlists.pms')
